requirements.txt
python-dotenv
mcp
pyodbc